[CGL] Moving landing page to separate file
