+++ .idea/dictionaries/p4003472.xml
@@ -1,0 +1,11 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="p4003472">
+    <words>
+      <w>httpd</w>
+      <w>inplace</w>
+      <w>pdfsandwich</w>
+      <w>pdftotext</w>
+      <w>tesseract</w>
+    </words>
+  </dictionary>
+</component>
