+++ cli-httpd/main.go
@@ -19,6 +19,11 @@ import (
 // To test it:
 // curl -X POST -H "Content-Type: application/json" -d '{"img_url":"http://localhost:8081/img","engine":0}' http://localhost:8081/ocr
 
+var (
+	sha1ver   string
+	buildTime string
+)
+
 func init() {
 	zerolog.TimeFieldFormat = time.StampMilli
 	// Default level is info, unless debug flag is present
@@ -53,6 +58,7 @@ func main() {
 
 	var httpPort uint
 	var debug bool
+	var flgVersion bool
 	flagFunc := func() {
 		flag.UintVar(
 			&httpPort,
@@ -66,48 +72,27 @@ func main() {
 			false,
 			"sets debug flag, program will print more messages",
 		)
+		flag.BoolVar(
+			&flgVersion,
+			"version",
+			false,
+			"show version and exit",
+		)
 	}
 
 	rabbitConfig := ocrworker.DefaultConfigFlagsOverride(flagFunc)
+	if flgVersion {
+		fmt.Printf("Build on %s from sha1 %s\n", buildTime, sha1ver)
+		os.Exit(0)
+	}
 	if debug {
 		zerolog.SetGlobalLevel(zerolog.DebugLevel)
 	}
 
 	// any requests to root, just redirect to main page
 	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		text := `<!DOCTYPE html><html lang="en"><head><meta charset="utf-8"><title>open-ocr</title>` +
-			`<style> html, body{font-family: "Fixedsys,Courier,monospace";}body {max-width: 960px; min-width: 320px;` +
-			`margin: 0 auto;}section {margin-top: 3em; margin: 3em 1.5em 0 1.5em;}section:last-of-type {margin-bottom: 3em;}` +
-			`li {margin-top: 0.8em;}.nes-container {position: relative; padding: 1.5rem 2rem; border-color: #000; border-style: solid;` +
-			`border-width: 4px;} .nes-container.with-title > .title {display: table;padding: 0 .5rem;margin: -2.2rem 0 1rem; font-size:` +
-			`1rem;background-color: #fff;}` +
-			`.nes-btn {border-image-slice: 2;border-image-width: 2;border-image-outset: 2;position: relative;border-style: solid;border-width: 4px;` +
-			`text-decoration: none;background-color: #92cc41;display: inline-block;padding: 6px 8px;color: #fff;` +
-			`border-image-source: url('data:image/svg+xml;utf8,<?xml version="1.0" encoding="UTF-8" ?><svg version="1.1" width="5" height="5" xmlns="http://www.w3.org/2000/svg"><path d="M2 1 h1 v1 h-1 z M1 2 h1 v1 h-1 z M3 2 h1 v1 h-1 z M2 3 h1 v1 h-1 z" fill="rgb(33,37,41)" /></svg>');}` +
-			`</style></head><body>` +
-			`<section class="nes-container with-title">	<h2 class="title">Open-ocr  ></h2>` +
-			`<div><a class="nes-btn is-success" href="">RUNNING</a>` +
-			`<pre>  __   ; '.'  :
-|  +. :  ;   :   __
-.-+\   ':  ;   ,,-'  :
-'.  '.  '\ ;  :'   ,'
-,-+'+. '. \.;.:  _,' '-.
-'.__  "-.::||::.:'___-+'
--"  '--..::::::_____.IIHb
-'-:______.:; '::,,...,;:HB\
-.+         \ ::,,...,;:HB \
-'-.______.:+ \'+.,...,;:P'  \
-.-'           \              \
-'-.______.:+   \______________\
-.-::::::::,     BBBBBBBBBBBBBBB
-::,,...,;:HB    BBBBBBBBBBBBBBB
-::,,...,;:HB    BBBBBBBBBBBBBBB
-::,,...,;:HB\   BBBBBBBBBBBBBBB
-::,,...,;:HB \  BBBBBBBBBBBBBBB
-'+.,...,;:P'  \ BBBBBBBBBBBBBBB
-               \BBBBBBBBBBBBBBB</pre><p>Nice day to put slinkies on an escalator!</p></div><p>proudly made with QBASIC:)</p>
-			<p>Need <a href="https://godoc.org/github.com/xf0e/open-ocr">docs</a>?</p></section></body> </html>`
-		fmt.Fprintf(w, text)
+		text := ocrworker.GenerateLandingPage()
+		_, _ = fmt.Fprintf(w, text)
 	})
 
 	//http.Handle("/ocr", ocrworker.NewOcrHttpHandler(rabbitConfig))
