--- prometheus.go
@@ -1,57 +1,0 @@
-package ocrworker
-
-import (
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"net/http"
-)
-
-// InstrumentHttpStatusHandler wraps httpHandler to provide prometheus metrics
-func InstrumentHttpStatusHandler(ocrHttpHandler *OcrHTTPStatusHandler) http.Handler {
-	inFlightGauge := prometheus.NewGauge(prometheus.GaugeOpts{
-		Name: "ocr_in_flight_requests",
-		Help: "A gauge of requests currently being served by the wrapped handler.",
-	})
-
-	counter := prometheus.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "ocr_api_requests_total",
-			Help: "A counter for requests to the wrapped handler.",
-		},
-		[]string{"code", "method"},
-	)
-
-	// duration is partitioned by the HTTP method and handler. It uses custom
-	// buckets based on the expected request duration.
-	duration := prometheus.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Name:    "ocr_request_duration_seconds",
-			Help:    "A histogram of latencies for requests.",
-			Buckets: []float64{.01, .05, .1, .25, .5, 1, 2.5, 5, 10},
-		},
-		[]string{"handler", "method"},
-	)
-
-	// requestSize has no labels, making it a zero-dimensional ObserverVec.
-	requestSize := prometheus.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Name:    "ocr_request_size_bytes",
-			Help:    "A histogram of response sizes for requests.",
-			Buckets: []float64{100, 1500, 5000000, 10000000, 25000000, 50000000},
-		},
-		[]string{},
-	)
-
-	// Register all of the metrics in the standard registry.
-	prometheus.MustRegister(inFlightGauge, counter, duration, requestSize)
-
-	ocrChain := promhttp.InstrumentHandlerInFlight(inFlightGauge,
-		promhttp.InstrumentHandlerDuration(duration.MustCurryWith(prometheus.Labels{"handler": "ocr"}),
-			promhttp.InstrumentHandlerCounter(counter,
-				//promhttp.InstrumentHandlerRequestSize(requestSize, ocrworker.NewOcrHttpHandler(rabbitConfig)),
-				promhttp.InstrumentHandlerRequestSize(requestSize, ocrHttpHandler),
-			),
-		),
-	)
-	return ocrChain
-}
